ChrisTheTrader/requirements.txt
python-dotenv>=1.0.0
ccxt>=4.4.0
pandas>=2.2.0
python-telegram-bot>=21.0
numpy>=1.26.0

